refactor(misc): improve compat, subtitle, watermark, ingest
- Improve compatibility filter handling
- Enhance CEA-608/708 subtitle processing
- Add watermark positioning improvements
- Improve RTMP ingest handling

+++ transcode-subtitle/src/cea708.rs
@@ -902,9 +902,9 @@ impl ServiceDecoder {
 
     /// Parse 6-bit color to Color (static version)
     fn parse_color_static(byte: u8) -> Color {
-        let r = ((byte & 0x03) * 85) as u8;
-        let g = (((byte >> 2) & 0x03) * 85) as u8;
-        let b = (((byte >> 4) & 0x03) * 85) as u8;
+        let r = (byte & 0x03) * 85;
+        let g = ((byte >> 2) & 0x03) * 85;
+        let b = ((byte >> 4) & 0x03) * 85;
         Color::rgb(r, g, b)
     }
